split links into sepparate files
